+++ Code/Domein/Toestel/IToestelRepository.cs
@@ -8,11 +8,11 @@ namespace Domein {
 
 		void VoegToestelToe(string naam);
 
-		void ZetToestelInOfUitHerstelling(long toestelId, bool nieuweHerstellingValue);
+		void ZetToestelInOfUitHerstelling(int toestelId, bool nieuweHerstellingValue);
 
 		void VerwijderToestel(Toestel toestel);
 
-		void UpdateToestelOpId(long toestelId, string toestelNaam);
-		bool GeefToestelHerstelStatusOpId(long toestelId);
+		void UpdateToestelOpId(int toestelId, string toestelNaam);
+		bool GeefToestelHerstelStatusOpId(int toestelId);
 	}
 }
